requirements.txt
flask==2.0.1
werkzeug==2.0.3
flask-cors==3.0.10
graphene==2.1.9
flask-graphql==2.0.1
pyjwt==2.1.0
bcrypt==3.2.0
python-dotenv==0.19.0
SQLAlchemy==1.4.23
graphql-core>=2.1,<3.0 